requirements/requirements.txt
boto3==1.34.51
python-dotenv==1.0.1
requests==2.31.0